+++ tests/testthat/test-provider-openai.R
@@ -222,3 +222,15 @@ test_that("value_turn() handles web_search_call action types", {
   )
   expect_equal(turn@contents[[1]]@query, "web search")
 })
+
+# Token counting -----------------------------------------------------------
+
+test_that("can count tokens", {
+  vcr::local_cassette("openai-count-tokens")
+  test_token_count(chat_openai_test)
+})
+
+test_that("can count tokens with tools", {
+  vcr::local_cassette("openai-count-tokens-tools")
+  test_token_count_tools(chat_openai_test)
+})